feat: implement ShopAdminPage and refactor context imports
