+++ Cardoon/src/components/pages/AdminPage/AdminPage.tsx
@@ -1,0 +1,147 @@
+import { useContext, useState, useEffect } from "react";
+import { UserContext } from "../../../context/UserContext";
+import { useFetch, RESOURCES, usePost } from "../../../hooks/server";
+import { Item } from "../../../types/common";
+
+export const ShopAdminPage = () => {
+  const { user } = useContext(UserContext);
+  const [newItem, setNewItem] = useState<Item>({
+    _id: "",
+    name: "Calculette Wish",
+    description: "Plus de gold, mais... êtes-vous sûr qu'elle fonctionne ?",
+    price: 500,
+    image:
+      "https://www.lissac.fr/media/catalog/product/0/3/03663234078213_ASH1907_NOIR_FACE_9131.png",
+    type: "weapon",
+    effect: {
+      type: "gold",
+      value: 5,
+    },
+  });
+  const [itemImageFile, setItemImageFile] = useState<File | string | null>(
+    null
+  );
+  const isDev = import.meta.env.MODE === "development";
+
+  const { fetch, data } = useFetch<Item[]>(RESOURCES.ITEMS);
+  const { post } = usePost<Item>(RESOURCES.ITEMS);
+  const [items, setItems] = useState<Item[]>(data || []);
+  useEffect(() => {
+    fetch();
+    document.title = "Page d'administration de la boutique";
+  }, []);
+
+  useEffect(() => {
+    if (data) {
+      setItems(data);
+    }
+  }, [data]);
+
+  if (!user || user.role !== "admin") {
+    return <p>Vous n'avez pas accès à cette page.</p>;
+  }
+
+  const saveItem = async (e: React.FormEvent) => {
+    e.preventDefault();
+    const formData = new FormData();
+    if (itemImageFile) {
+      formData.append("imageFile", itemImageFile);
+    }
+    formData.append("image", newItem.image);
+    formData.append("name", newItem.name);
+    formData.append("description", newItem.description);
+    formData.append("price", newItem.price.toString());
+    formData.append("type", newItem.type);
+    formData.append("effectType", newItem.effect.type);
+    formData.append("effectValue", newItem.effect.value.toString());
+
+    await post(formData);
+  };
+
+  if (!isDev) {
+    return <p>Cette page est réservée aux administrateurs.</p>;
+  }
+
+  return (
+    <div className="Admin">
+      <h1>Page d'administration de la boutique</h1>
+      <p>Cette page est réservée aux administrateurs.</p>
+      <h2>Ajouter un nouvel objet</h2>
+      <form onSubmit={saveItem}>
+        <input
+          type="text"
+          placeholder="Nom de l'objet"
+          value={newItem.name}
+          onChange={(e) => setNewItem({ ...newItem, name: e.target.value })}
+        />
+        <input
+          type="text"
+          placeholder="Description de l'objet"
+          value={newItem.description}
+          onChange={(e) =>
+            setNewItem({ ...newItem, description: e.target.value })
+          }
+        />
+        <input
+          type="number"
+          placeholder="Prix de l'objet"
+          value={newItem.price}
+          onChange={(e) =>
+            setNewItem({ ...newItem, price: parseInt(e.target.value) })
+          }
+        />
+        <input
+          type="text"
+          placeholder="Image de l'objet (URL)"
+          value={newItem.image}
+          onChange={(e) => setNewItem({ ...newItem, image: e.target.value })}
+        />
+        <input
+          type="file"
+          accept="image/*"
+          onChange={(e) => {
+            if (e.target.files && e.target.files[0]) {
+              setItemImageFile(e.target.files[0]);
+            }
+          }}
+        />
+        <select
+          value={newItem.type}
+          onChange={(e) =>
+            setNewItem({ ...newItem, type: e.target.value as Item["type"] })
+          }
+        >
+          <option value="head">Tête</option>
+          <option value="weapon">Arme</option>
+          <option value="armor">Armure</option>
+          <option value="accessory">Accessoire</option>
+        </select>
+        <input
+          type="number"
+          placeholder="Valeur de l'effet"
+          value={newItem.effect.value}
+          onChange={(e) =>
+            setNewItem({
+              ...newItem,
+              effect: { ...newItem.effect, value: parseInt(e.target.value) },
+            })
+          }
+        />
+        <button type="submit">Ajouter</button>
+      </form>
+      <h2>Liste des objets</h2>
+      <ul>
+        {items.map((item) => (
+          <li key={item._id}>
+            <h3>{item.name}</h3>
+            <p>{item.description}</p>
+            <p>Prix: {item.price} gold</p>
+            <img src={item.image} alt={item.name} />
+          </li>
+        ))}
+      </ul>
+    </div>
+  );
+};
+
+export default ShopAdminPage;
